fixed warning: type of ‘mpi_allreduce’
changed definition of size variable to kind=int64 in dust_mod.f90

+++ source/dust_mod.f90
@@ -25,7 +25,7 @@ module dust_mod
 
       integer :: err ! allocation error status
       integer :: iP,jP,kP ! counter
-      integer :: size ! size of MPI reducing string
+      integer(kind=int64) :: size ! size of MPI reducing string
       integer :: yTop
 
       logical,save :: lgFirstDustEm=.true.
